+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* Added expert option to disable background upload/download.
+
+### Changed
+* Client: if there are annotations, confirm before closing/cancelling
+
+### Fixed
+
+
+## [0.3.0] - 2020-02-10
+
+### Added
 * saved comments are filtered per question and per test.
 * Marker has a "view" button to look at any test number.
 
@@ -17,8 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Client: config and saved comments now more human-readable in toml files.
 * Client: can download test/server info before logging in.
 * Client is more pessimistic about errors and will crash instead of trying to continue
-in most cases.	
+in most cases.
 * Client checks for double-logins and can force logout a previous session.
+* Client: you must make at least one annotation on the page.
 
 ### Fixed
 * many fixes, especially related to client crashes due to networking.
@@ -108,7 +120,8 @@ in most cases.
 
 This is the first release of Plom, PaperLessOpenMarking.
 
-[Unreleased]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.2.2...master
+[Unreleased]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.3.0...master
+[0.3.0]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.2.2...v0.3.0
 [0.2.2]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.2.1...v0.2.2
 [0.2.1]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.2.0...v0.2.1
 [0.2.0]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.1.0...v0.2.0